+++ lib/pave/obbit.tcl
@@ -46,44 +46,44 @@ namespace eval ::apave {
   set ::apave::_CS_(ALL) {
 {{ 0: AwLight} "#141414" #151616 #dfdfde #d1d1d0 #562222 #85b4e7 #000 #444 grey #1a497c #000 #7fb8ce - #bebebd #000 #FBFB96 #cacaca #a20000 #76b2f1 #005 #006 #007}
 {{ 1: AzureLight} "#050b0d" #050b0d #fff #e1e1e1 #002aaa #8dd9db #000 #444 grey #0e7b7c #000 #81cdcf - #cccccc #000 #FBFB95 #e2e2e0 #ad0000 #76b2f1 #005 #006 #007}
-{{ 2: ForestLight} "#050b0d" #050b0d #fff #e1e1e1 #004000 #A8CCA8 #000 #444 grey #217346 #000 #a8bda8 - #cccccc #000 #FBFB95 #e2e2e0 #ad0000 #76b2f1 #005 #006 #007}
+{{ 2: ForestLight} "#050b0d" #050b0d #fff #e1e1e1 #004000 #A8CCA8 #000 #185818 grey #217346 #000 #a8bda8 - #cccccc #000 #FBFB95 #e2e2e0 #ad0000 #76b2f1 #005 #006 #007}
 {{ 3: SunValleyLight} "#050b0d" #050b0d #fff #e1e1e1 #00469f #74c9ff #000 #444 grey #005fb8 #000 #7fcbff - #cccccc #000 #FBFB95 #e2e2e0 #950000 #76b2f1 #005 #006 #007}
 {{ 4: Grey1} "#050b0d" #050b0d #F8F8F8 #dadad8 #5c1616 #AFAFAF #000 #444 grey #933232 #000 #AFAFAF - #caccd0 #000 #FBFB95 #e0e0d8 #a20000 #76b2f1 #005 #006 #007}
 {{ 5: Grey2} "#050b0d" #050b0d #e9e9e7 #F8F8F8 #5c1616 #b8b8b8 #000 #444 grey #933232 #000 #c1c1c1 - #e1e1e1 #000 #FBFB95 #d5d5d3 #a20000 #76b2f1 #005 #006 #007}
-{{ 6: Rosy} "#2B122A" #000 #FFFFFF #F6E6E9 #570957 #C5ADC8 #000 #444 grey #870287 #000 #ceb6d1 - #e3d3d6 #000 #FBFB95 #e5e3e1 #a20000 #76b2f1 #005 #006 #007}
+{{ 6: Rosy} "#2B122A" #000 #FFFFFF #F6E6E9 #570957 #C5ADC8 #000 #630063 grey #870287 #000 #ceb6d1 - #e3d3d6 #000 #FBFB95 #e5e3e1 #a20000 #76b2f1 #005 #006 #007}
 {{ 7: Clay} "#000" #000 #fdf4ed #ded3cc #500a0a #bcaea2 #000 #444 grey #843500 #fff #9a8f83 - #d5c9c1 #000 #FBFB95 #e1dfde #a20000 #76b2f1 #005 #006 #007}
-{{ 8: Dawn} "#08085D" #030358 #FFFFFF #e3f9f9 #562222 #a3dce5 #000 #444 grey #933232 #000 #99d2db - #d3e9e9 #000 #FBFB96 #dbe9ed #a20000 #76b2f1 #005 #006 #007}
-{{ 9: Sky} "#102433" #0A1D33 #d0fdff #bdf6ff #562222 #95ced7 #000 #444 grey #933232 #000 #9ad3dc - #b1eaf3 #000 #FBFB95 #c0e9ef #a20000 #76b2f1 #005 #006 #007}
-{{10: Florid} "#000" #004000 #e4fce4 #fff #5c1616 #93e493 #0F2D0F #444 grey #802e00 #004000 #a7f8a7 - #d8e7d8 #000 #FBFB96 #d7e6d7 #a20000 #76b2f1 #005 #006 #007}
-{{11: LightGreen} "#122B05" #091900 #edffed #DEF8DE #562222 #A8CCA8 #000 #444 grey #933232 #000 #A8CCA8 - #cde7cd #000 #FBFB96 #dee9de #a20000 #76b2f1 #005 #006 #007}
-{{12: InverseGreen} "#122B05" #091900 #cce6c8 #DEF8DE #562222 #9cc09c #000 #444 grey #933232 #000 #b5d9b5 - #c9e3c9 #000 #FBFB96 #bed8ba #a20000 #76b2f1 #005 #006 #007}
-{{13: GreenPeace} "#001000" #001000 #e1ffdd #cadfca #562222 #9dbb99 #000 #444 grey #933232 #000 #9cb694 - #b9d3b9 #000 #FBFB96 #d2e1d2 #a20000 #76b2f1 #005 #006 #007}
-{{14: African} "#000" #000 #fff #ffffe7 #460000 #ffd797 #000 #6f2509 #7e7e7e #771d00 #000 #e6ae80 - #e7e7cf #000 #eded89 #ededd5 #a20000 #76b2f1 #005 #006 #007}
-{{15: African1} "#000" #000 #f5f5dd #f2ebd2 #460000 #ffc48a #000 #6f2509 #7e7e7e #771d00 #000 #f2ebd2 - #e3dcc3 #000 #eded89 #e3e3cb #a20000 #76b2f1 #005 #006 #007}
-{{16: African2} "#000" #000 #ffffe4 #eae7c0 #500a0a #eaac7a #000 #6f2509 grey #771d00 #00003c #e6ae80 - #dddab3 #000 #fbfb74 #e7e7cb #a20000 #76b2f1 #005 #006 #007}
-{{17: African3} "#000" #000 #fdf9d0 #d5d2af #500a0a #d59d6f #000 #6f2509 grey #771d00 #00003c #e6ae80 - #c5c29f #000 #fbfb74 #e5e5cc #c10000 #76b2f1 #005 #006 #007}
-{{18: Yellowstone} "#00002f" #00003c #ffffd1 #cfcdb1 #591c0e #c89160 #000 #444 grey #771d00 #3b1516 #cfab86 - #c2c0a4 #000 #ffff45 #e6e6bb #a30000 #76b2f1 #005 #006 #007}
-{{19: Notebook} "#000" #000 #e9e1c8 #c2bca8 #460000 #d59d6f #000 #444 #7e7e7e #771d00 #000 #c09c77 - #d0cab6 #000 #eded89 #dad2b9 #a20000 #76b2f1 #005 #006 #007}
-{{20: Notebook1} "#000" #000 #dad2b9 #b5af9b #460000 #d59d6f #000 #444 #707070 #771d00 #000 #ba9671 - #c5bfab #000 #eded89 #ccc4ab #a20000 #76b2f1 #005 #006 #007}
-{{21: Notebook2} "#000" #000 #cdc5ac #a6a08c #460000 #d59d6f #000 #444 #606060 #771d00 #000 #cfab86 - #b4ae9a #000 #eded89 #c1b9a0 #980000 #76b2f1 #005 #006 #007}
-{{22: Notebook3} "#000" #000 #beb69d #96907c #460000 #d59d6f #000 #444 #505050 #771d00 #000 #cfab86 - #a6a08c #000 #eded89 #b2aa91 #7b1010 #76b2f1 #005 #006 #007}
+{{ 8: Dawn} "#08085D" #030358 #FFFFFF #e3f9f9 #562222 #a3dce5 #000 #195999 grey #933232 #000 #99d2db - #d3e9e9 #000 #FBFB96 #dbe9ed #a20000 #76b2f1 #005 #006 #007}
+{{ 9: Sky} "#102433" #0A1D33 #d0fdff #bdf6ff #562222 #95ced7 #000 #195999 grey #933232 #000 #9ad3dc - #b1eaf3 #000 #FBFB95 #c0e9ef #a20000 #76b2f1 #005 #006 #007}
+{{10: Florid} "#000" #004000 #e4fce4 #fff #5c1616 #93e493 #0F2D0F #185818 grey #802e00 #004000 #a7f8a7 - #d8e7d8 #000 #FBFB96 #d7e6d7 #a20000 #76b2f1 #005 #006 #007}
+{{11: LightGreen} "#122B05" #091900 #edffed #DEF8DE #562222 #A8CCA8 #000 #185818 grey #933232 #000 #A8CCA8 - #cde7cd #000 #FBFB96 #dee9de #a20000 #76b2f1 #005 #006 #007}
+{{12: InverseGreen} "#122B05" #091900 #cce6c8 #DEF8DE #562222 #9cc09c #000 #185818 grey #933232 #000 #b5d9b5 - #c9e3c9 #000 #FBFB96 #bed8ba #a20000 #76b2f1 #005 #006 #007}
+{{13: GreenPeace} "#001000" #001000 #e1ffdd #cadfca #562222 #9dbb99 #000 #185818 grey #933232 #000 #9cb694 - #b9d3b9 #000 #FBFB96 #d2e1d2 #a20000 #76b2f1 #005 #006 #007}
+{{14: African} "#000" #000 #fff #ffffe7 #460000 #ffd797 #000 #682800 #7e7e7e #771d00 #000 #e6ae80 - #e7e7cf #000 #eded89 #ededd5 #a20000 #76b2f1 #005 #006 #007}
+{{15: African1} "#000" #000 #f5f5dd #f2ebd2 #460000 #ffc48a #000 #682800 #7e7e7e #771d00 #000 #f2ebd2 - #e3dcc3 #000 #eded89 #e3e3cb #a20000 #76b2f1 #005 #006 #007}
+{{16: African2} "#000" #000 #ffffe4 #eae7c0 #500a0a #eaac7a #000 #682800 grey #771d00 #00003c #e6ae80 - #dddab3 #000 #fbfb74 #e7e7cb #a20000 #76b2f1 #005 #006 #007}
+{{17: African3} "#000" #000 #fdf9d0 #d5d2af #500a0a #d59d6f #000 #682800 grey #771d00 #00003c #e6ae80 - #c5c29f #000 #fbfb74 #e5e5cc #c10000 #76b2f1 #005 #006 #007}
+{{18: Yellowstone} "#00002f" #00003c #ffffd1 #cfcdb1 #591c0e #c89160 #000 #682800 grey #771d00 #3b1516 #cfab86 - #c2c0a4 #000 #ffff45 #e6e6bb #a30000 #76b2f1 #005 #006 #007}
+{{19: Notebook} "#000" #000 #e9e1c8 #c2bca8 #460000 #d59d6f #000 #682800 #7e7e7e #771d00 #000 #c09c77 - #d0cab6 #000 #eded89 #dad2b9 #a20000 #76b2f1 #005 #006 #007}
+{{20: Notebook1} "#000" #000 #dad2b9 #b5af9b #460000 #d59d6f #000 #682800 #707070 #771d00 #000 #ba9671 - #c5bfab #000 #eded89 #ccc4ab #a20000 #76b2f1 #005 #006 #007}
+{{21: Notebook2} "#000" #000 #cdc5ac #a6a08c #460000 #d59d6f #000 #682800 #606060 #771d00 #000 #cfab86 - #b4ae9a #000 #eded89 #c1b9a0 #980000 #76b2f1 #005 #006 #007}
+{{22: Notebook3} "#000" #000 #beb69d #96907c #460000 #d59d6f #000 #682800 #505050 #771d00 #000 #cfab86 - #a6a08c #000 #eded89 #b2aa91 #7b1010 #76b2f1 #005 #006 #007}
 {{23: Darcula} "#ececec" #c7c7c7 #272727 #323232 #e98f1c #2F5692 #e1e1e1 #f4f49f grey #d18d3f #EDC881 #2a518d - #444444 #000 #a2a23e #343434 #f28787 #76b2f1 #005 #006 #007}
 {{24: Dusk} "#ececec" #ececec #1a1f21 #262b2d #95bf95 #217346 #FFF #f4f49f #585d5f #99c399 #ffffff #225c3b - #363b3d #000 #9e9e3a #23282a #ffabab #99dd99 #005 #006 #007}
 {{25: AwDark} "#F0E8E8" #E7E7E7 #1f2223 #232829 #de9e5e #215d9c #fff #f4f49f grey #80bcfb #fff #134f8e - #313637 #000 #cdcd69 #292e2f #ffabab #76b2f1 #005 #006 #007}
-{{26: AzureDark} "#ececec" #c7c7c7 #272727 #393939 #28a7ff #007fff #FFF #f4f49f grey #33b2ff #EDC881 #0062a5 - #4a4a4a #000 #a2a23e #404040 #ffc341 #76b2f1 #005 #006 #007}
-{{27: ForestDark} "#ececec" #c7c7c7 #272727 #393939 #95bf95 #217346 #FFF #f4f49f grey #99c399 #ffffff #247649 - #4a4a4a #000 #a2a23e #323232 #efaf6f #99dd99 #005 #006 #007}
-{{28: SunValleyDark} "#dfdfdf" #dddddd #131313 #252525 #38a9e0 #2f60d8 #FFF #f4f49f #6f6f6f #57c8ff #fff #2051c9 - #323232 #000 #a2a23e #2a2a2a #efaf6f #4273eb #005 #006 #007}
+{{26: AzureDark} "#ececec" #c7c7c7 #272727 #393939 #28a7ff #0a89c1 #FFF #f4f49f grey #33b2ff #EDC881 #0062a5 - #4a4a4a #000 #a2a23e #383838 #ffc341 #76b2f1 #005 #006 #007}
+{{27: ForestDark} "#ececec" #c7c7c7 #272727 #393939 #95bf95 #217346 #FFF #42ff42 grey #99c399 #ffffff #247649 - #4a4a4a #000 #a2a23e #383838 #efaf6f #99dd99 #005 #006 #007}
+{{28: SunValleyDark} "#dfdfdf" #dddddd #131313 #252525 #38a9e0 #2a627f #FFF #f4f49f #6f6f6f #57c8ff #fff #2051c9 - #323232 #000 #a2a23e #2a2a2a #efaf6f #4273eb #005 #006 #007}
 {{29: Dark1} "#E0D9D9" #C4C4C4 #212121 #292929 #de9e5e #6c6c6c #000 #f4f49f #606060 #eda95b #000 #767676 - #363636 #000 #cdcd69 #292929 #ffabab #76b2f1 #005 #006 #007}
 {{30: Dark2} "#bebebe" #bebebe #1f1f1f #262626 #de9e5e #6b6b6b #000 #f4f49f #616161 #eda95b #000 #767676 - #323232 #000 #b0b04c #262626 #ffabab #76b2f1 #005 #006 #007}
 {{31: Dark3} "#bebebe" #bebebe #0a0a0a #232323 #de9e5e #6a6a6a #000 #f4f49f #616161 #eda95b #000 #767676 - #303030 #000 #bebe5a #131313 #ffabab #76b2f1 #005 #006 #007}
-{{32: Oscuro} "#f1f1f1" #f1f1f1 #344545 #526d6d #f1b479 #728d8d #fff #f4f49f #afafaf #f7c475 #000 #94afaf - #5d7474 #000 #cdcd69 #3d4e4e #ffbcbc #a3f1c7 #005 #006 #007}
-{{33: Oscuro1} "#f1f1f1" #f1f1f1 #2a3b3b #466161 #e5a565 #6c8787 #fff #f4f49f #a2a2a2 #edba6b #000 #8ba6a6 - #526969 #000 #cdcd69 #354646 #ffbcbc #96e4ba #005 #006 #007}
-{{34: Oscuro2} "#f1f1f1" #f1f1f1 #223333 #3e5959 #de9e5e #668181 #fff #f4f49f #a2a2a2 #e4b162 #000 #819c9c - #4c6363 #000 #cdcd69 #2b3c3c #ffbcbc #8ad8ae #005 #006 #007}
-{{35: Oscuro3} "#f1f1f1" #f1f1f1 #192a2a #355050 #de9e5e #5c7777 #fff #f4f49f #9e9e9e #dfac5d #000 #779292 - #425959 #000 #cdcd69 #223333 #ffbcbc #80cea4 #005 #006 #007}
-{{36: MildDark} "#d2d2d2" #fff #222323 #384e66 #2ccaca #4b7391 #fff #00ffff #939393 #43e1e1 #000 #668eac - #455b73 #000 #bebe5a #2b2c2c #ffa2a2 #76b2f1 #005 #006 #007}
-{{37: MildDark1} "#d2d2d2" #fff #151616 #2D435B #2ac8c8 #436b89 #fff #00ffff grey #36d4d4 #000 #668eac - #3a5068 #000 #bebe5a #1f2020 #ffb0b0 #76b2f1 #005 #006 #007}
-{{38: MildDark2} "#b4b4b4" #fff #0d0e0e #24384f #28c6c6 #3e6684 #fff #00ffff #757575 #33d1d1 #000 #668eac - #33475e #000 #bebe5a #161717 #ffaeae #76b2f1 #005 #006 #007}
-{{39: MildDark3} "#e2e2e2" #f1f1f1 #000 #1B3048 #27c5c5 #375f7d #fff #00ffff #6c6c6c #31d0d0 #000 #668eac - #2a3f57 #000 #b0b04c #0f0f0f #ffafaf #76b2f1 #005 #006 #007}
+{{32: Oscuro} "#f1f1f1" #ffffff #314242 #3e5959 #f1b479 #6c8787 #fff #42ff42 #afafaf #f7c475 #000 #94afaf - #4d6868 #000 #cbcb67 #425353 #ffcd8d #94e2b8 #005 #006 #007}
+{{33: Oscuro1} "#e3e3e3" #f7f7f7 #233434 #304b4b #e3a66b #5e7979 #fff #42ff42 #a1a1a1 #e9b667 #000 #86a1a1 - #3f5a5a #000 #bfbf5b #344545 #ffc585 #86d4aa #005 #006 #007}
+{{34: Oscuro2} "#d5d5d5" #f1f1f1 #152626 #223d3d #d5985d #506b6b #fff #42ff42 #939393 #dba859 #000 #789393 - #314c4c #000 #bbbb57 #263737 #f7bc7c #78c69c #005 #006 #007}
+{{35: Oscuro3} "#c7c7c7" #eaeaea #071818 #142f2f #c78a4f #425d5d #fff #42ff42 #858585 #cd9a4b #000 #6a8585 - #233e3e #000 #afaf4b #182929 #e9ae6e #6ab88e #005 #006 #007}
+{{36: MildDark} "#d2d2d2" #ffffff #223142 #2D435B #2ac8c8 #517997 #fff #00ffff grey #36d4d4 #000 #668eac - #3a5068 #000 #bebe5a #324152 #eeae6e #76b2f1 #005 #006 #007}
+{{37: MildDark1} "#c8c8c8" #f7f7f7 #1a2937 #24384f #28c6c6 #466e8c #fff #00ffff #757575 #33d1d1 #000 #668eac - #31455c #000 #bebe5a #2b3a48 #f1b171 #76b2f1 #005 #006 #007}
+{{38: MildDark2} "#e2e2e2" #f1f1f1 #0e1d2c #1B3048 #27c5c5 #426a88 #fff #00ffff #6c6c6c #31d0d0 #000 #668eac - #2a3f57 #000 #b0b04c #1d2c3b #ebab6b #76b2f1 #005 #006 #007}
+{{39: MildDark3} "#dbdbdb" #eaeaea #000c1b #031830 #27c5c5 #375f7d #fff #00ffff #6c6c6c #28c7c7 #000 #5a82a0 - #162b43 #000 #b0b04c #0a1f37 #e5a565 #76b2f1 #005 #006 #007}
 {{40: Inkpot} "#d3d3ff" #AFC2FF #16161f #1E1E27 #de9e5e #6767a8 #000 #f4f49f #6e6e6e #ffbb6d #000 #8585c6 - #292936 #000 #a2a23e #202029 #ffa5a5 #7a7abb #005 #006 #007}
 {{41: Quiverly} "#cdd8d8" #cdd8d8 #2b303b #333946 #de9e5e #6f7582 #000 #f4f49f #757575 #eda95b #000 #9197a4 - #414650 #000 #b0b04c #323742 #ffabab #76b2f1 #005 #006 #007}
 {{42: Monokai} "#f8f8f2" #f8f8f2 #353630 #4e5044 #f1b479 #707070 #000 #f4f49f #9a9a9a #ffbb6d #000 #777777 - #46473d #000 #cdcd69 #3c3d37 #ffabab #cd994b #005 #006 #007}
@@ -95,6 +95,7 @@ namespace eval ::apave {
 }
   set ::apave::_CS_(initall) 1
   set ::apave::_CS_(initWM) 1
+  set ::apave::_CS_(isActive) 1
   set ::apave::_CS_(!FG) #000000
   set ::apave::_CS_(!BG) #b7b7b7 ;#a8bcd2 #c3c3c3 #9cb0c6 #4a6984
   set ::apave::_CS_(expo,tfg1) "-"
@@ -293,6 +294,18 @@ proc ::apave::initWM {args} {
   initPOP .
   initStyles
   return
+}
+
+#########################################################################
+
+proc ::apave::cs_Active {{flag ""}} {
+
+  # Gets/sets "is changing CS possible" flag for a whole application.
+
+  if {[string is boolean -strict $flag]} {
+    set ::apave::_CS_(isActive) $flag
+  }
+  return $::apave::_CS_(isActive)
 }
 
 #########################################################################
@@ -1423,6 +1436,10 @@ oo::class create ::apave::ObjectTheming {
     # E.g., in TKE editor, e_menu and add_shortcuts plugins use it to
     # be consistent with TKE theme.
 
+    if {![::apave::cs_Active]} {
+      my themeMandatory $win {*}$args
+      return
+    }
     lassign $clrs tfg1 tbg1 tfg2 tbg2 tfgS tbgS tfgD tbgD tcur bclr \
       thlp tfgI tbgI tfgM tbgM twfg twbg tHL2 tbHL chkHL res5 res6 res7
     if {$tfg1 eq {-}} return
@@ -1645,6 +1662,18 @@ oo::class create ::apave::ObjectTheming {
       set ::apave::_C_($ts,1) "-foreground $tfg1"
       set ::apave::_C_($ts,2) "-background $tbg1"
     }
+    my themeMandatory $win {*}$args
+    return
+  }
+
+  ###########################################################################
+
+  method themeMandatory {win args} {
+
+    # Themes all that must be themed.
+    #   win - window's name
+    #   args - options
+
     # set the new options for nested widgets (menu e.g.)
     my themeNonThemed $win
     # other options per widget type